axi_adc_trigger: Add cascade support.
- Add embedded trigger option. This will enable the use of the
embedded trigger as an option in the data stream for further
processing, keeping the data synchronized with the trigger.
When module cascading(synchronization) is desired
(logic_analyzer - adc_trigger) a small propagation time is
required, hence the need to remove the util_extract(trigger extract)
module from the data path is a must.

- Add more options for the IO triggering. This will open the door
for multiple M2k synchronizations.